0419/backend/services/face_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
google-generativeai>=0.3.1
python-dotenv==1.0.0
